fix broken single mode transaction and bug fix of multi mode transaction

+++ src/dialogue/add_transaction/mod.rs
@@ -9,7 +9,7 @@ pub enum AddTransactionState {
     Idle,
     Started,
     AmountAsked {
-        amount: i64,
+        amount: f64,
     },
 }
